backend/requirements.txt
# backend/requirements.txt - CORRECTED VERSIONS

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI & LangChain
langchain==0.1.0
langchain-groq==0.1.0
langchain-chroma==0.1.2  # Changed from 0.1.0 to 0.1.2 (available version)
langchain-huggingface==0.0.3

# Vector Store & Embeddings
chromadb==0.4.22
sentence-transformers==2.2.2

# PDF Processing
pymupdf==1.23.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0